added highscore component
